python3-sense-hat: update recipe to use python3
Recipe in meta-raspberrypi layer was moved to use python3 and use pypi to
download and install it, however this pypi package does not contain the
examples files provided in the original source code.

This commit maintains the original source code file, to allow installing the
example files, but modifies the bbappend recipe to use python3 binary.

https://jira.digi.com/browse/DEL-7159

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>